+++ lib/framework.dart
@@ -13,5 +13,6 @@ library framework;
 export 'src/framework/buffers.dart';
 export 'src/framework/converters.dart';
 export 'src/framework/dart/buffers.dart';
+export 'src/framework/dart/filters.dart';
 export 'src/framework/filters.dart';
 export 'src/framework/sinks.dart';